update engines field to match reality
async-listener mostly works fine with Node.js 0.12.0 and io.js 1.x;
the old engines field was included under the assumption that
async-listener was going to be a polyfill compatible with
`process.{add,create,remove}Listener` in node core. That's not how
things played out, and people are using async-listener in projects on
newer Nodes, so ditch the warning.