Allow failures on latest Node 7
Because issue https://github.com/othiym23/async-listener/issues/109 have
not yet been solved for Node.js 7, the test suite currently fails when
running on the latest version of Node.js 7 (v7.10.0 as of this commit).

This commit allows 7.x to fail, but adds an extra build for 7.9.0 which
is know to be good.

This should be reverted when/if Node.js 7 is patched accordingly.